requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
haystack-ai==2.0.0
openai==1.12.0
numpy==1.26.4
joblib==1.3.2
python-multipart==0.0.9

# Development dependencies
pre-commit==3.6.0
black==24.2.0
isort==5.13.2
mypy==1.8.0
